# specs/001-it-inventory/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: IT Inventory Management System API
  description: RESTful API for managing IT equipment inventory, assignments, and reporting
  version: 1.0.0
  contact:
    name: IT Department
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://inventory.company.com/api
    description: Production server
...
paths:
...
  /items:
    get:
      tags: [Items]
      summary: List inventory items
      operationId: listItems
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
          description: Search in name, serial, asset ID, brand, model
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ItemCategory'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ItemStatus'
        - name: condition
          in: query
          schema:
            $ref: '#/components/schemas/ItemCondition'
        - name: location
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, assetId, createdAt, purchaseDate, warrantyExpiration]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Items]
      summary: Create new item
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /items/{id}:
    get:
      tags: [Items]
      summary: Get item by ID
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Items]
      summary: Update item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Items]
      summary: Soft delete item (decommission)
      operationId: softDeleteItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Item decommissioned
        '400':
          description: Cannot delete assigned item
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /items/{id}/status:
    patch:
      tags: [Items]
      summary: Change item status manually
      operationId: changeItemStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, reason]
              properties:
                status:
                  $ref: '#/components/schemas/ItemStatus'
                reason:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /items/bulk-delete:
    post:
      tags: [Items]
      summary: Bulk soft delete items
      operationId: bulkDeleteItems
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemIds, reason]
              properties:
                itemIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                reason:
                  type: string
      responses:
        '200':
          description: Bulk delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                  skipped:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        reason:
                          type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ItemId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ItemCategory:
      type: string
      enum: [LAPTOP, MONITOR, KEYBOARD, MOUSE, HEADSET, PHONE, TABLET, CABLES, SOFTWARE_LICENSE, OTHER]
    ItemStatus:
      type: string
      enum: [AVAILABLE, ASSIGNED, RESERVED, IN_REPAIR, DECOMMISSIONED, LOST_STOLEN]
    ItemCondition:
      type: string
      enum: [NEW, GOOD, FAIR, NEEDS_REPAIR, DECOMMISSIONED]

    # ========== USER ==========
...
    ItemSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetId:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/ItemCategory'
        status:
          $ref: '#/components/schemas/ItemStatus'
        condition:
          $ref: '#/components/schemas/ItemCondition'
        brand:
          type: string
        model:
          type: string
        thumbnailUrl:
          type: string
        currentAssignee:
          type: string
          description: Employee name if assigned
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemSummary'
        - type: object
          properties:
            serialNumber:
              type: string
            purchaseDate:
              type: string
              format: date
            purchasePrice:
              type: number
            warrantyExpiration:
              type: string
              format: date
            location:
              type: string
            notes:
              type: string
            photoUrl:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            createdBy:
              type: string
              description: User name
    CreateItemRequest:
      type: object
      required: [name, category, serialNumber]
      properties:
        assetId:
          type: string
          description: Auto-generated if not provided
        name:
          type: string
          maxLength: 200
        category:
          $ref: '#/components/schemas/ItemCategory'
        condition:
          $ref: '#/components/schemas/ItemCondition'
        brand:
          type: string
        model:
          type: string
        serialNumber:
          type: string
          maxLength: 100
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          minimum: 0
        warrantyExpiration:
          type: string
          format: date
        location:
          type: string
        notes:
          type: string
        photo:
          type: string
          format: binary
    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          $ref: '#/components/schemas/ItemCategory'
        condition:
          $ref: '#/components/schemas/ItemCondition'
        brand:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
        warrantyExpiration:
          type: string
          format: date
        location:
          type: string
        notes:
          type: string
        photo:
          type: string
          format: binary

    # ========== EMPLOYEE ==========
...
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
        total:
          type: integer

    # Constitution Art. 4.2 mandates this error format
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code (e.g., VALIDATION_ERROR)
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string